+++ jannovar-hgvs/src/main/java/de/charite/compbio/jannovar/hgvs/protein/ProteinDeletion.java
@@ -1,6 +1,6 @@
-package de.charite.compbio.jannovar.hgvs.change.protein;
+package de.charite.compbio.jannovar.hgvs.protein;
 
-import de.charite.compbio.jannovar.hgvs.change.AminoAcidCode;
+import de.charite.compbio.jannovar.hgvs.AminoAcidCode;
 
 // TODO(holtgrew): ProteinSeqSpecification should also store whether to display at all and we should integrate this here
