+++ jannovar-hgvs/src/main/java/de/charite/compbio/jannovar/hgvs/protein/ProteinSilentChange.java
@@ -1,6 +1,6 @@
-package de.charite.compbio.jannovar.hgvs.change.protein;
+package de.charite.compbio.jannovar.hgvs.protein;
 
-import de.charite.compbio.jannovar.hgvs.change.AminoAcidCode;
+import de.charite.compbio.jannovar.hgvs.AminoAcidCode;
 
 /**
  * Represents a silent protein-level change, i.e., "p.=".
