+++ jannovar-hgvs/src/test/java/de/charite/compbio/jannovar/hgvs/protein/ProteinSilentChangeTest.java
@@ -1,9 +1,10 @@
-package de.charite.compbio.jannovar.hgvs.change.protein;
+package de.charite.compbio.jannovar.hgvs.protein;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-import de.charite.compbio.jannovar.hgvs.change.AminoAcidCode;
+import de.charite.compbio.jannovar.hgvs.AminoAcidCode;
+import de.charite.compbio.jannovar.hgvs.protein.ProteinSilentChange;
 
 public class ProteinSilentChangeTest {
